Move detectOnPath to its own file for now
